LayoutTests/http/tests/webrtc/enumerateDevicesInFrames-expected.txt

CONSOLE MESSAGE: Permission policy 'Camera' check failed for document with origin 'http://127.0.0.1:8000'.
CONSOLE MESSAGE: Permission policy 'Microphone' check failed for document with origin 'http://127.0.0.1:8000'.
CONSOLE MESSAGE: Not allowed to call enumerateDevices.


PASS allow attribute can be updated dynamically
